src/requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.17.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
babel==2.16.0
bcrypt==4.1.2
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.2
fastadmin==0.2.13
fastapi==0.112.1
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.15.4
flake8==7.1.1
flake8-broken-line==1.0.0
flake8-docstrings==1.7.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.6.0
idna==3.8
iniconfig==2.0.0
iso8601==1.1.0
isort==5.13.2
Jinja2==3.1.4
makefun==1.15.4
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pendulum==3.0.0
pep8-naming==0.14.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pwdlib==0.2.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.30
time-machine==2.15.0
tortoise-orm==0.21.6
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
virtualenv==20.26.3
watchfiles==0.23.0
websockets==13.0
